build: Fix build warnings with GCC 10
Also,
* Fix cyclical dependency between bdb and berkdb due to os_namemangle.c
* Clean up dead-code dealing with later versions for berkdb
* Fix a few false-positive warnings (string-overflow and return-local-addr)
