+++ bdb/attr.c
@@ -119,13 +119,8 @@ static void bdb_attr_set_int(bdb_state_type *bdb_state, bdb_attr_type *bdb_attr,
         if (bdb_state) {
             if (bdb_state->dbenv) {
                 int rc;
-#if defined(BERKDB_4_5) || defined(BERKDB_46)
-                rc = bdb_state->dbenv->rep_set_limit(bdb_state->dbenv, 0,
-                                                     bdb_attr->replimit);
-#else
                 rc = bdb_state->dbenv->set_rep_limit(bdb_state->dbenv, 0,
                                                      bdb_attr->replimit);
-#endif
                 if (rc)
                     logmsg(LOGMSG_ERROR, "%s:set_rep_limit: %d %s\n", __func__,
                            rc, bdb_strerror(rc));
