+++ bdb/file.c
@@ -1287,9 +1287,6 @@ bdb_state_type *gbl_bdb_state;
 
 char *bdb_trans(const char infile[], char outfile[])
 {
-#ifdef COMPILING_FOR_DB_TOOLS
-    return strcpy(outfile, infile);
-#else
     bdb_state_type *bdb_state;
     char *p;
     int len;
@@ -1324,10 +1321,7 @@ char *bdb_trans(const char infile[], char outfile[])
         memmove(p, p + 4, len - 4 + 1 /* copy \0 byte too! */);
     }
 
-    /*printf("bdb_trans: <%s> -> <%s>\n", infile, outfile);*/
-
     return outfile;
-#endif
 }
 
 int net_hostdown_rtn(netinfo_type *netinfo_ptr, char *host);
@@ -2362,14 +2356,6 @@ static DB_ENV *dbenv_open(bdb_state_type *bdb_state)
         return NULL;
     }
 
-#ifdef BERKDB_46
-    rc = dbenv->rep_set_timeout(dbenv, DB_REP_CHECKPOINT_DELAY, 0);
-    if (rc != 0) {
-        logmsg(LOGMSG_FATAL, "__rep_set_timeout failed\n");
-        exit(1);
-    }
-#endif
-
     rc = dbenv->set_paniccall(dbenv, panic_func);
 
 #ifdef BDB_VERB_REPLICATION_DEFAULT
@@ -2560,11 +2546,6 @@ static DB_ENV *dbenv_open(bdb_state_type *bdb_state)
 
     flags = DB_CREATE | DB_PRIVATE | DB_THREAD | DB_INIT_LOCK | DB_INIT_LOG |
             DB_INIT_MPOOL | DB_INIT_TXN | DB_INIT_REP;
-
-#if defined(BERKDB_4_5) || defined(BERKDB_46)
-    /* register the berkdb event callback handler */
-    dbenv->set_event_notify(dbenv, berkdb_event_func);
-#endif
 
     /* register the routine that berkeley db will use to send data
        over the network */
@@ -2885,11 +2866,7 @@ if (!is_real_netinfo(bdb_state->repinfo->netinfo))
    */
 
 /* limit amount we retrans in one shot */
-#if defined(BERKDB_4_5) || defined(BERKDB_46)
-    rc = dbenv->rep_set_limit(dbenv, 0, bdb_state->attr->replimit);
-#else
     rc = dbenv->set_rep_limit(dbenv, 0, bdb_state->attr->replimit);
-#endif
     if (rc != 0) {
         logmsg(LOGMSG_FATAL, "%s: dbenv->set_rep_limit(%u) %d %s\n", __func__,
                 bdb_state->attr->replimit, rc, bdb_strerror(rc));
@@ -3021,103 +2998,7 @@ waitformaster:
               lsn_to_str(our_lsn_str, &our_lsn));
     }
 
-/*goto done;*/
-
-/*
-   PHASE 1:
-   wait until berkdb claims we are "caught up"
-   */
-
-/* berkdb 4.2 doesnt support startup done message, so skip this phase */
-#if defined(BERKDB_4_3) || defined(BERKDB_4_5) || defined(BERKDB_46)
-
-again1:
-    if (!gbl_skip_catchup_logic && bdb_state->repinfo->master_host != myhost) {
-        master_host = bdb_state->repinfo->master_host;
-        if (master_host == myhost)
-            goto done1;
-        if ((master_host == db_eid_invalid) || (master_host == bdb_master_dupe))
-            goto waitformaster;
-
-        if ((master_host < 0) || (master_host > (MAXNODES - 1))) {
-            logmsg(LOGMSG_FATAL, "master_host == %s\n", master_sid);
-            exit(1);
-        }
-
-        bzero(&(bdb_state->seqnum_info->seqnums[nodeix(myhost)]),
-              sizeof(seqnum_type));
-
-        memcpy(&master_seqnum,
-               &(bdb_state->seqnum_info->seqnums[nodeix(master_sid)]),
-               sizeof(seqnum_type));
-        memcpy(&master_lsn, &master_seqnum.lsn, sizeof(DB_LSN));
-
-        /*
-        fprintf(stder, "%s:%d master_seqnum=%d:%d\n", __FILE__, __LINE__,
-              master_lsn.file, master_lsn.offset);
-         */
-
-        rc = bdb_state->dbenv->log_stat(bdb_state->dbenv, &log_stats,
-                                        DB_STAT_VERIFY);
-        if (rc != 0) {
-            logmsg(LOGMSG_INFO, "err %d from log_stat\n", rc);
-            goto again1;
-        }
-
-        make_lsn(&our_lsn, log_stats->st_cur_file, log_stats->st_cur_offset);
-        free(log_stats);
-
-        if (count > 10) {
-            /*
-               if we're in the middle of processing a message, back off
-               from whining and try again the next second.  goal is not to
-               commitdelaymore the master when we're actually in a long
-               rep_process_message. yes, we can race, but better than
-               nothing.
-            */
-            if (bdb_state->repinfo->in_rep_process_message) {
-                count--;
-                sleep(1);
-                goto again1;
-            }
-
-/* i dont think we need to do this anymore. */
-#if 0         
-         rc = net_send_message(bdb_state->repinfo->netinfo, 
-            master_eid,
-            USER_TYPE_ADD_DUMMY,
-            &master_eid, sizeof(int), 0, 0);
-         if (rc != 0)
-         {
-            fprintf(stderr, "net_send to %d failed rc %d\n", 
-               master_eid, rc);
-            exit(1);
-         }
-#endif
-
-            rc = print_catchup_message(bdb_state, 1, &our_lsn, &master_lsn,
-                                       &gap, &prev_gap, &catchup_state,
-                                       starting_time, &starting_lsn);
-            if (rc != 0) {
-                goto again1;
-            }
-
-            count = 0;
-        }
-        count++;
-
-        if (!bdb_state->berkdb_rep_startupdone) {
-            sleep(1);
-            goto again1;
-        }
-    }
-done1:
-    logmsg(LOGMSG_DEBUG, "phase 1 replication catchup passed\n");
-
-#endif
-
     /*
-       PHASE 2:
        wait until _IN REALITY_ the lsn we have is pretty darn close to
        the lsn of the master
        */
@@ -3135,11 +3016,6 @@ again2:
                &(bdb_state->seqnum_info->seqnums[nodeix(master_host)]),
                sizeof(seqnum_type));
         memcpy(&master_lsn, &master_seqnum.lsn, sizeof(DB_LSN));
-
-        /*
-        fprintf(stderr, "%s:%d master_seqnum=%d:%d\n", __FILE__, __LINE__,
-              master_lsn.file, master_lsn.offset);
-         */
 
         rc = bdb_state->dbenv->log_stat(bdb_state->dbenv, &log_stats, 0);
         if (rc != 0) {
@@ -3163,20 +3039,6 @@ again2:
                 sleep(1);
                 goto again2;
             }
-
-/* i dont think we need to do this anymore. */
-#if 0
-         rc = net_send_message(bdb_state->repinfo->netinfo, 
-            bdb_state->repinfo->master_eid,
-            USER_TYPE_ADD_DUMMY,
-            &bdb_state->repinfo->master_eid, sizeof(int), 0, 0);
-         if (rc != 0)
-         {
-            fprintf(stderr, "net_send to %d failed rc %d\n", 
-               bdb_state->repinfo->master_eid, rc);
-            exit(1);
-         }
-#endif
 
             rc = print_catchup_message(bdb_state, 1, &our_lsn, &master_lsn,
                                        &gap, &prev_gap, &catchup_state,
@@ -3206,7 +3068,6 @@ done2:
     logmsg(LOGMSG_DEBUG, "phase 2 replication catchup passed\n");
 
     /*
-       PHASE 3:
        Tell the master where we are.  he knows where he is.  when he determines
        we are "close enough" (defined by "delta") then we pass phase 3
        and go into syncronous mode.
