+++ berkdb/db/db_iface.c
@@ -391,9 +391,7 @@ __db_cursor_pp_paired(dbp, dbc_old, dbcp, flags)
 	return rc;
 }
 
-#ifndef BERKDB_46
 pthread_key_t DBG_FREE_CURSOR;
-#endif
 
 int
 transfer_cursor_ownership(DBC *cold, DBC *cnew)
@@ -723,7 +721,6 @@ __db_cursor_real(dbp, txn, txn_clone, dbcp_old, lid_clone, dbcs, dbcp, flags,
 #endif
 
 		if (countmein) {
-#ifndef BERKDB_46
 			struct __dbg_free_cursor *p = pthread_getspecific(DBG_FREE_CURSOR);
 			if (p) {
 				p->counter++;
@@ -738,7 +735,6 @@ __db_cursor_real(dbp, txn, txn_clone, dbcp_old, lid_clone, dbcs, dbcp, flags,
 					Pthread_setspecific(DBG_FREE_CURSOR, p);
 				}
 			}
-#endif
 		}
 	}
