+++ db/comdb2.h
@@ -2482,8 +2482,6 @@ int has_index_changed(struct dbtable *db, char *keynm, int ct_check, int newkey,
 int resume_schema_change(void);
 
 void debug_trap(char *line, int lline);
-int count_db(struct dbtable *db);
-int compact_db(struct dbtable *db, int timeout, int freefs);
 int ix_find_last_dup_rnum_kl(struct ireq *iq, int ixnum, void *key, int keylen,
                              void *fndkey, int *fndrrn,
                              unsigned long long *genid, void *fnddta,
