+++ tools/cdb2_printlog/cdb2_printlog.c
@@ -106,9 +106,7 @@ int tool_cdb2_printlog_main(argc, argv)
 	comdb2ma_init(0, 0);
 
 	Pthread_key_create(&comdb2_open_key, NULL);
-#ifndef BERKDB_46
 	Pthread_key_create(&DBG_FREE_CURSOR, NULL);
-#endif
 
 	if ((ret = cdb2_print_version_check(progname)) != 0)
 		return (ret);
